sinkNode v 0.8 - debugging process
sinkNode working based on command of the base node. In this version there is update of MAC address for sink node 1 & 2 due to the hardware issue. Sink node become smarter whenever receive the command from base node it will do re-checking roomTemp of the server with the current AC temp. The system will only change the temp whenever the msg.roomTemp is on the determined threshold (for raising or lowering temp). If, the AC is broken then, the sink node will command the AC to be set at 16 or 30deg based on temp change in server room